00-pitch: Add Python string vs bytearray demo
The demo highlights the need to understand language datatypes and
language runtime internals. In this case Python strings versus the
bytearray data type. Python strings are immutable, so update / append
operations will imply the creation of a new string; the byterray data
type, on the other hand, is mutable and can be modified in place - a
faster operation than on a string.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>